+++ src/scene/shader-lib/chunks/gsplat/vert/gsplatOutput.js
